feat(MDwrapper): wrap listObject operations
We use an ugly hack to get back bucketUIDs when needed, they will be
dropped soon enough, cf #116.

- Add listObject function in the wrapper
- Remove listObject methods in Bucket
- Update the listObject API to have more encapsulation
- Remove calls to said methods
- Replace them by the wrapper function
- Update associated unit tests
