+++ lib/storage/metadata/in_memory/Bucket.js
@@ -1,12 +1,11 @@
 import naturalCompare from 'natural-compare-lite';
 
-import { markerFilter, markerFilterMPU,
-    prefixFilter, findNextMarker } from './bucket_utilities';
-import { ListBucketResult } from './ListBucketResult';
-import { ListMultipartUploadsResult } from './ListMultipartUploadsResult';
 import config from '../../../config';
-const splitter = config.splitter;
+import { markerFilterMPU, prefixFilter, findNextMarker }
+    from './bucket_utilities';
+import { ListMultipartUploadsResult } from './ListMultipartUploadsResult';
 
+const splitter = config.splitter;
 const defaultMaxKeys = 1000;
 
 export default class Bucket {
@@ -24,88 +23,6 @@ export default class Bucket {
         this.name = name;
         this.owner = owner;
         this.creationDate = new Date;
-    }
-
-    getBucketListObjects(
-        prefix, marker, delimiter, paramMaxKeys = defaultMaxKeys, callback) {
-        if (prefix && typeof prefix !== 'string') {
-            return callback('InvalidArgument');
-        }
-
-        if (marker && typeof marker !== 'string') {
-            return callback('InvalidArgument');
-        }
-
-        if (delimiter && typeof delimiter !== 'string') {
-            return callback('InvalidArgument');
-        }
-
-        if (paramMaxKeys && typeof paramMaxKeys !== 'number') {
-            return callback('InvalidArgument');
-        }
-
-        // If paramMaxKeys is undefined, the default parameter will set it.
-        // However, if it is null, the default parameter will not set it.
-        let maxKeys = paramMaxKeys;
-        if (maxKeys === null) {
-            maxKeys = defaultMaxKeys;
-        }
-
-        const response = new ListBucketResult();
-        let keys = Object.keys(this.keyMap).sort(naturalCompare);
-        // If marker specified, edit the keys array so it
-        // only contains keys that occur alphabetically after the marker
-        if (marker) {
-            keys = markerFilter(marker, keys);
-            response.Marker = marker;
-        }
-        // If prefix specified, edit the keys array so it only
-        // contains keys that contain the prefix
-        if (prefix) {
-            keys = prefixFilter(prefix, keys);
-            response.Prefix = prefix;
-        }
-        // Iterate through keys array and filter keys containing delimiter
-        // into response.CommonPrefixes and filter remaining
-        // keys into response.Contents
-        for (let i = 0; i < keys.length; i++) {
-            const currentKey = keys[i];
-            // Do not list object with delete markers
-            if (response.hasDeleteMarker(currentKey, this.keyMap) === true) {
-                continue;
-            }
-            // If hit maxKeys, stop adding keys to response
-            if (response.MaxKeys >= maxKeys) {
-                response.IsTruncated = true;
-                response.NextMarker = findNextMarker(i, keys, response);
-                break;
-            }
-            // If a delimiter is specified, find its
-            // index in the current key AFTER THE OCCURRENCE OF THE PREFIX
-            let delimiterIndexAfterPrefix = -1;
-            let prefixLength = 0;
-            if (prefix) {
-                prefixLength = prefix.length;
-            }
-            const currentKeyWithoutPrefix = currentKey.slice(prefixLength);
-            let sliceEnd;
-            if (delimiter) {
-                delimiterIndexAfterPrefix = currentKeyWithoutPrefix
-                    .indexOf(delimiter);
-                sliceEnd = delimiterIndexAfterPrefix + prefixLength;
-                response.Delimiter = delimiter;
-            }
-            // If delimiter occurs in current key, add key to
-            // response.CommonPrefixes.
-            // Otherwise add key to response.Contents
-            if (delimiterIndexAfterPrefix > -1) {
-                const keySubstring = currentKey.slice(0, sliceEnd + 1);
-                response.addCommonPrefix(keySubstring);
-            } else {
-                response.addContentsKey(currentKey, this.keyMap);
-            }
-        }
-        return callback(null, response);
     }
 
     getMultipartUploadListing(params, callback) {
