+++ lib/storage/metadata/in_memory/ListBucketResult.js
@@ -9,15 +9,17 @@ export class ListBucketResult extends ListResult {
     addContentsKey(key, keyMap) {
         const objectMetadata = keyMap[key];
         this.Contents.push({
-            Key: decodeURIComponent(key),
-            LastModified: objectMetadata['last-modified'],
-            ETag: objectMetadata['content-md5'],
-            StorageClass: objectMetadata['x-amz-storage-class'],
-            Owner: {
-                DisplayName: objectMetadata['owner-display-name'],
-                ID: objectMetadata['owner-id']
+            key: decodeURIComponent(key),
+            value: {
+                LastModified: objectMetadata['last-modified'],
+                ETag: objectMetadata['content-md5'],
+                StorageClass: objectMetadata['x-amz-storage-class'],
+                Owner: {
+                    DisplayName: objectMetadata['owner-display-name'],
+                    ID: objectMetadata['owner-id']
+                },
+                Size: objectMetadata['content-length']
             },
-            Size: objectMetadata['content-length']
         });
         this.MaxKeys += 1;
     }
