+++ lib/storage/metadata/in_memory/bucket_utilities.js
@@ -86,5 +86,5 @@ export function findNextMarker(currentIndex, array, responseObject) {
 }
 
 export function isKeyInContents(responseObject, key) {
-    return responseObject.Contents.some(val => val.Key === key);
+    return responseObject.Contents.some(val => val.key === key);
 }
